requirements.txt
# Django
Django==5.1.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.5.*
django-cors-headers==4.6.*

# FastAPI
fastapi==0.115.*
uvicorn[standard]==0.30.*

# DB (PostgreSQL)
psycopg[binary]==3.2.*

# Config
python-dotenv==1.0.*
pydantic==2.9.*

# Auth (2FA, JWT)
pyotp==2.9.*
Pillow==11.0.*

# Dev tooling (buenas prácticas)
ruff==0.6.*
black==24.8.*
mypy==1.11.*
pytest==8.3.*
pytest-django==4.9.*
httpx==0.27.*

# Optional: para CORS en FastAPI
starlette==0.41.*
